+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 # Unreleased
 
+- Add `ocaml.jump` to jump to a specific target. (#1654)
+
 ## 1.22.0
 
 - Fix formatting of cwd path on windows (#1650)
 - Add `ocaml.construct` to construct an expression from a typedhole. (#1646)
-- Add `ocaml.jump` to jump to a specific target. (#1654)
 
 ## 1.21.0